Added a class (ImageProc) to connect to and pull images from the database. This class will eventually read the reference colors from the database too. RepColors still completely ignores ImageProc, and will happily read arguments in from the command-line. Argument processing is not terribly robust, but should work with correctly specified arguments.
As of this commit, this project officially requires JDK 1.7 or higher.

Cleaned up the code a bit to deal with warnings.

+++ maketest.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-rm colors.html
-for i in $(cat colors.txt); do echo '<div class="foo" style="background-color:rgb('$i')"></div>' >> colors.html; done
+[[ -e colors.html ]] && rm colors.html
+for i in $(cat colors.csv); do echo '<div class="foo" style="background-color:rgb('$i')"></div>' >> colors.html; done
 cat testheader.html colors.html > fulltest.html
 echo '</body></html>' >> fulltest.html
